+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Full documentation for rocSOLVER is available at [rocsolver.readthedocs.io](http
     - GESVDJ (with batched and strided\_batched versions)
 - LU factorization without pivoting for block tridiagonal matrices:
     - GEBLTTRF_NPVT (with batched, strided\_batched, and interleaved\_batched versions)
+- Linear system solver without pivoting for block tridiagonal matrices:
+    - GEBLTTRS_NPVT (with batched, strided\_batched, and interleaved\_batched versions)
 
 ### Optimized
 - Improved the performance of SYEVJ/HEEVJ.